# Speed up circuit construction when contents are a list of moments